% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluspect.R
\name{define.bands}
\alias{define.bands}
\title{define.bands}
\usage{
define.bands()
}
\value{
a spectral object.
}
\description{
\code{define.bands} defines the spectral regions for the Fluspect-B model
}
\details{
Define spectral regions for SCOPE v_1.40 \cr
All spectral regions are defined here as row vectors \cr
WV Jan. 2013
}
\examples{
spectral <- define.bands()

}
\author{
Nastassia Vilfan, Christiaan van der Tol, Onno Muller, Uwe Rascher, Wouter Verhoef (Original version in Matlab)

Alberto Hornero (Ported version into R)
}
